add authorization for users

+++ src/tasks/tasks.service.ts
@@ -5,6 +5,7 @@ import { TaskRepository } from './task.repository';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Task } from './task.entity';
 import { TaskStatus } from './task-status.enum';
+import { User } from 'src/auth/user.entity';
 
 @Injectable()
 export class TasksService {
@@ -13,31 +14,31 @@ export class TasksService {
     private taskRepository: TaskRepository,
   ) {}
 
-  public async getTasks(filterDto: GetTasksFilterDto): Promise<Task[]> {
-    return this.taskRepository.getTasks(filterDto);
+  public async getTasks(filterDto: GetTasksFilterDto, user: User): Promise<Task[]> {
+    return this.taskRepository.getTasks(filterDto, user);
   }
 
-  public async createTask(createTaskDto: CreateTaskDto): Promise<Task> {
-    return this.taskRepository.createTask(createTaskDto);
+  public async createTask(createTaskDto: CreateTaskDto, user: User): Promise<Task> {
+    return this.taskRepository.createTask(createTaskDto, user);
   }
 
-  public async getTaskById(id: number): Promise<Task> {
-    const task = await this.taskRepository.findOne(id);
+  public async getTaskById(id: number, user: User): Promise<Task> {
+    const task = await this.taskRepository.findOne({ where: { id, userId: user.id } });
     if (!task) {
       throw new NotFoundException(`Task with id "${id}" not found.`);
     }
     return task;
   }
 
-  public async deleteTask(id: number): Promise<void> {
-    const res = await this.taskRepository.delete(id);
+  public async deleteTask(id: number, user: User): Promise<void> {
+    const res = await this.taskRepository.delete({ id, userId: user.id });
     if (res.affected === 0) {
       throw new NotFoundException(`Task with id "${id}" not found.`); 
     }
   }
 
-  public async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-    const task = await this.getTaskById(id);
+  public async updateTaskStatus(id: number, status: TaskStatus, user: User): Promise<Task> {
+    const task = await this.getTaskById(id, user);
     task.status = status;
     await task.save();
     return task;
